Rename file snippetvamp.php to index.php
